Move functions into seperate modules and refactor
* Move convert_to_dataframe and parse_contents into utilities, rename them and refactor them.
* Create new trace component and move relevant code to it, refactor.
* Move all callbacks from create_graph to its controller, refactor and fix type hint errors.
